requirements.txt
# Core dependencies
click>=8.1.7
pyyaml>=6.0.1
rich>=13.7.0
openai>=1.12.0
python-dotenv>=1.0.0
jsonschema>=4.21.0
requests>=2.31.0

# API dependencies (for Docker deployment)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
redis>=5.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
pydantic-settings>=2.1.0

# Sigma Translator dependencies
ruamel.yaml>=0.18.5
httpx>=0.26.0
